Issue #196: Migrating to java.time (JSR-310). First cut.
- Deleted deprecated methods.

- Base unit of time is now nanosecond unlike millisecond in Joda.

- Correspondence between Joda classes and Java:
DateTime        ZonedDateTime
DateTimeZone    ZoneId, ZoneOffset
LocalDate       LocalDate
LocalDateTime   LocalDateTime
LocalTime       LocalTime
Period          Duration, Period
ReadablePeriod  TemporalAmount

- No Java equivalents of Joda classes:
DateMidnight

- Joda Interval class is replaced by Interval record.

- I commented out methods which don't have unit tests, don't have obvious
replacement in Java, or their behaviour in Java is different than in Joda.
Some of them may be implemented in the next cut.

- I tried to preserve as many unit tests as possible and, based on the
result, it seems that in most cases the new implementation is a drop-in
replacement of Joda.